updated Dashboard report and system
