--- src/Controller/Admin/DashboardSystemCrudController.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace App\Controller\Admin;
-
-use App\Entity\System;
-use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
-use EasyCorp\Bundle\EasyAdminBundle\Field\IdField;
-use EasyCorp\Bundle\EasyAdminBundle\Field\TextEditorField;
-use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
-
-class DashboardSystemCrudController extends AbstractCrudController
-{
-    public static function getEntityFqcn(): string
-    {
-        return System::class;
-    }
-
-    /*
-    public function configureFields(string $pageName): iterable
-    {
-        return [
-            IdField::new('id'),
-            TextField::new('title'),
-            TextEditorField::new('description'),
-        ];
-    }
-    */
-}
