+++ src/Controller/Admin/ThemeCrudController.php
@@ -6,6 +6,7 @@ use App\Entity\Theme;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use EasyCorp\Bundle\EasyAdminBundle\Field\IdField;
+use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
 
 class ThemeCrudController extends AbstractCrudController
 {
@@ -18,16 +19,18 @@ class ThemeCrudController extends AbstractCrudController
     public function configureFields(string $pageName): iterable
     {
         $id = IdField::new('id');
+        $name = TextField::new('name');
+
 
 
         if (Crud::PAGE_INDEX === $pageName) {
-            return [$id];
+            return [$name];
         }
         elseif(Crud::PAGE_DETAIL === $pageName) {
-            return [];
+            return [$id];
         }
         elseif(Crud::PAGE_NEW === $pageName) {
-            return [];
+            return [$id];
         }
 //        elseif(Crud::PAGE_EDIT === $pageName) {
 //            return [$title, $editor_description, $editor_instructions, $editor_preparations, $coll_questions, $coll_configuration ];
